+++ src/segment/nu-segment.cpp
@@ -242,7 +242,7 @@ int main (int argc, char **argv){
 #endif
 		}
 		t2 = cciutils::ClockGetTime();
-		printf("%d::%d: segment %d us, in %s\n", rank, tid, (int)(t2-t1), fin.c_str());
+		printf("%d::%d: segment %lu us, in %s\n", rank, tid, t2-t1, fin.c_str());
 //		std::cout << rank <<"::" << tid << ":" << t2-t1 << " us, in " << fin << ", out " << fmask << std::endl;
 
 
@@ -286,7 +286,7 @@ int main (int argc, char **argv){
 #endif
     if (rank == 0)  {
     	t4 = cciutils::ClockGetTime();
-		printf("**** Segment took %d us\n", (int)(t4-t3));
+		printf("**** Segment took %lu us\n", t4-t3);
 	//	std::cout << "**** Segment took " << t4-t3 << " us" << std::endl;
 
 		t3 = cciutils::ClockGetTime();
